// src/models/product.ts
export interface ApiProduct {
    id: number;
    title: string;
    description: string;
    price: number;
    category: string;
    thumbnail: string;
    stock: number;
    discountPercentage: number;
  }
  
  export interface Product {
    id: number;
    title: string;
    description: string;
    price: string;
    category: string;
    thumbnail: string;
    stock: number;
    discountPercentage: number;
  }
  